Error in IP address
Corrected the following line where IP addresses contained an error in which the 1 was dropped from the last octet. 

Original text:
We know the latter 445 port is typically SMB, so we can perhaps make an assumption that 192.168.1.57 is the target machine with it’s SMB port exposed. This makes 192.168.1.49 the attacker machine.

Corrected text:

We know the latter 445 port is typically SMB, so we can perhaps make an assumption that 192.168.1.157 is the target machine with it’s SMB port exposed. This makes 192.168.1.149 the attacker machine.

Great write up but the way!